--- Web/src/main/java/User/UserRepository.java
@@ -1,9 +1,0 @@
-package User;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<SiteUser, Long>{
-	Optional<SiteUser> findByusername(String sname);
-}
