--- Web/src/main/java/User/UserService.java
@@ -1,37 +1,0 @@
-package User;
-
-import java.util.Optional;
-
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import com.example.demo.DataNotFoundException;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Service
-public class UserService {
-
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    public SiteUser create(String sname, String sid, String semail, String password) {
-        SiteUser user = new SiteUser();
-        user.setUsername(sname);
-        user.setId(sid);
-        user.setEmail(email);
-        user.setPassword(passwordEncoder.encode(password));
-        this.userRepository.save(user);
-        return user;
-    }
-    
-    public SiteUser getUser(String sname) {
-        Optional<SiteUser> siteUser = this.userRepository.findByusername(sname);
-        if (siteUser.isPresent()) {
-            return siteUser.get();
-        } else {
-            throw new DataNotFoundException("siteuser not found");
-        }
-    }
-}
