+++ Web/src/main/java/com/example/demo/user/SiteUser.java
@@ -1,4 +1,4 @@
-package User;
+package com.example.demo.user;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
