+++ Web/src/main/java/com/example/demo/user/UserRepository.java
@@ -1,5 +1,9 @@
 package com.example.demo.user;
 
-public class UserRepository {
+import java.util.Optional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<SiteUser, Long>{
+	Optional<SiteUser> findByusername(String sname);
 }
