+++ Web/src/main/java/com/example/demo/user/UserRole.java
@@ -1,4 +1,4 @@
-package User;
+package com.example.demo.user;
 
 import lombok.Getter;
