+++ Web/src/main/java/com/example/demo/user/UserSecurityService.java
@@ -1,4 +1,4 @@
-package User;
+package com.example.demo.user;
 
 import java.util.ArrayList;
 import java.util.List;
